category and order controller

+++ app/Http/Controllers/CategoryController.php
@@ -25,26 +25,33 @@ class CategoryController extends Controller
         return response()->json($category, 201);
     }
 
-    public function show(Category $category)
+    public function show ($id)
     {
-        return response()->json($category);
+        $category = Category::findOrFail($id);
+        return response()->json($category, 200);
     }
 
-    public function update(Request $request, Category $category)
+    public function update(Request $request, $id)
     {
         $request->validate([
             'category_name' => 'required',
         ]);
-
+        $category = Category::findOrFail($id);
         $category->update($request->all());
 
         return response()->json($category, 200);
     }
 
-    public function destroy(Category $category)
+    public function destroy($id)
     {
+        $category = Category::find($id);
+    
+        if (!$category) {
+            return response()->json(['message' => 'Category not found'], 404);
+        }
+    
         $category->delete();
-
+    
         return response()->json(null, 204);
     }
 }
